refactor(scale): move scale and hx711 under hardware
