+++ src/libs/business_logic/business_logic.py
@@ -12,8 +12,8 @@ from libs.ui import userinterface as ui_module
 from libs.data import data_interface as data_module
 
 from RPi.GPIO import GPIO
-from libs.scale.scale_hardware.tatobari_hx711.hx711 import HX711
-from libs.scale.scale import Scale
+from libs.hardware.tatobari_hx711.hx711 import HX711
+from libs.hardware.scale.scale import Scale
 
 from libs.hardware.timing_calculator.calculator import Calculator
 from serial import Serial
