--- src/libs/scale/scale_hardware/__init__.py
@@ -1,2 +1,0 @@
-# tatobari_hx711.hx711.HX711 for real world usage
-# tatobari_hx711.emulated_hx711. for unit tests without hardware
